1.Fix the build error caused by the modification for compiler 6. 2.Revise startup_NUC230_240.s and retarget.c files to fix build issue. 3.Modify settings of Keil project to speed up compilation. 4.Revise copyright format for gen. chm file. 5.Fix SPI_Loopback GCC sample code issue.

+++ SampleCode/RegBased/I2C_Wakeup_Slave/main.c
@@ -6,8 +6,9 @@
  * @brief    NUC230_240 Series I2C Driver Sample Code for Wake-up (Slave)
  *
  * @note
- * Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
+ * @copyright SPDX-License-Identifier: Apache-2.0
  *
+ * @copyright Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
  ******************************************************************************/
 #include <stdio.h>
 #include "NUC230_240.h"
@@ -349,7 +350,7 @@ int32_t main(void)
     __NOP();
 
     while((g_u8SlvPWRDNWK & g_u8SlvI2CWK) == 0);		
-    printf("Power-down Wake-up INT 0x%x\n", ((CLK->PWRCON) & CLK_PWRCON_PD_WU_STS_Msk));		
+    printf("Power-down Wake-up INT 0x%x\n", (unsigned int)((CLK->PWRCON) & CLK_PWRCON_PD_WU_STS_Msk));		
     printf("I2C0 WAKE INT 0x%x\n", I2C0->I2CWKUPSTS);
